refactor: update imports to use capabilities module instead of ai-capabilities
