+++ src/app/drawers/SettingsDrawer/panels/LanguageSettings.tsx
@@ -7,9 +7,9 @@ import MenuItem from "@mui/material/MenuItem";
 import Select from "@mui/material/Select";
 import Stack from "@mui/material/Stack";
 import Typography from "@mui/material/Typography";
+import { isTranslatorSupported } from "@shared/ai/capabilities";
 import { useAI } from "@shared/ai/useAI";
 import { useLanguage } from "@shared/language/useLanguage";
-import { isTranslatorSupported } from "@shared/ai/ai-capabilities";
 
 export function LanguageSettings() {
   const { languages, languageCode, setLanguageCode } = useLanguage();
