+++ src/shared/ai/AICapabilitiesList.tsx
@@ -8,7 +8,7 @@ import {
   isProofreaderSupported,
   isRewriterSupported,
   isTranslatorSupported,
-} from "./ai-capabilities";
+} from "./capabilities";
 
 const AI_FEATURES = [
   { isSupported: isProofreaderSupported, label: "Proofreader" },
