+++ src/shared/ai/capabilities.ts
